+++ ruuvitag_sensor/ruuvi_types.py
@@ -2,7 +2,7 @@ from typing import Optional, Tuple, TypedDict, Union
 
 
 class SensorDataBase(TypedDict):
-    data_format: int
+    data_format: int | str
 
 
 class SensorDataUrl(SensorDataBase):
@@ -53,6 +53,12 @@ class SensorData6(SensorDataBase):
     mac: str
 
 
+class SensorDataE1(SensorData6):
+    pm_1: float
+    pm_4: float
+    pm_10: float
+
+
 class SensorHistoryData(TypedDict):
     humidity: Optional[float]
     temperature: Optional[float]
@@ -62,7 +68,7 @@ class SensorHistoryData(TypedDict):
 
 SensorData = Union[SensorDataUrl, SensorData3, SensorData5, SensorData6]
 
-DataFormat = Optional[int]
+DataFormat = Optional[int | str]
 RawSensorData = Optional[str]
 DataFormatAndRawSensorData = Tuple[DataFormat, RawSensorData]