+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### [Unreleased]
 * CHANGE: Support only Python 3.10 and newer
-* ADD: Support for Ruuvi Air (Data Format 6)
+* ADD: Support for Ruuvi Air (Data Formats 6 and E1)
 
 ## [3.1.0] - 2025-03-01
 * ADD: Support for fetching history data from RuuviTags